some little bug fixes in op_funcs and battle loop

+++ srcs/vm/funcs/0x02_ld.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   0x02_ld.c                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: orantane <orantane@student.hive.fi>        +#+  +:+       +#+        */
+/*   By: seronen <seronen@student.hive.fi>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/02/03 13:37:30 by seronen           #+#    #+#             */
-/*   Updated: 2021/02/09 15:35:57 by orantane         ###   ########.fr       */
+/*   Updated: 2021/02/09 20:32:49 by seronen          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,7 +22,7 @@ void    op_ld(t_carriage *carriage, unsigned char *arena)
 		carriage->regs[reg_slot] = carriage->stmt->args[0];
 	else if (carriage->stmt->arg_types[0] == T_IND)
 	{
-		arg_idx_value = (short)carriage->stmt->args[0] % IDX_MOD;
+		arg_idx_value = carriage->stmt->args[0] % IDX_MOD;
 		carriage->regs[reg_slot] = read_bytes(arena, (carriage->abs_pos + arg_idx_value) % IDX_MOD, 4);
 	}
 	if (carriage->regs[reg_slot] == 0)
