+++ srcs/vm/funcs/0x03_st.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   0x03_st.c                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: orantane <orantane@student.hive.fi>        +#+  +:+       +#+        */
+/*   By: seronen <seronen@student.hive.fi>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/02/05 15:48:06 by Gerhard           #+#    #+#             */
-/*   Updated: 2021/02/09 15:35:58 by orantane         ###   ########.fr       */
+/*   Updated: 2021/02/09 20:32:42 by seronen          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,7 +26,7 @@ void    op_st(t_carriage *carriage, unsigned char *arena)
 		carriage->regs[arg_2 - 1] = arg_value;
 	else if (carriage->stmt->arg_types[1] == T_IND)
 	{
-		arg_idx_value = (short)arg_2 % IDX_MOD;
+		arg_idx_value = arg_2 % IDX_MOD;
 		write_bytes(arena, (carriage->abs_pos + arg_idx_value) % IDX_MOD, arg_value, 4);
 	}
 }
