+++ srcs/vm/funcs/0x09_zjmp.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   0x09_zjmp.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: orantane <orantane@student.hive.fi>        +#+  +:+       +#+        */
+/*   By: seronen <seronen@student.hive.fi>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/02/06 15:02:48 by seronen           #+#    #+#             */
-/*   Updated: 2021/02/09 15:36:12 by orantane         ###   ########.fr       */
+/*   Updated: 2021/02/09 20:32:31 by seronen          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,7 +18,7 @@ void    op_zjmp(t_carriage *carriage, unsigned char *arena)
 
 	if (carriage->carry_flag)
 	{
-		arg_idx_value = (short)carriage->stmt->args[0] % IDX_MOD;
+		arg_idx_value = carriage->stmt->args[0] % IDX_MOD;
 		carriage->pos = fetch_position(arena, carriage, arg_idx_value);
 		carriage->abs_pos += arg_idx_value - 3;
 	}
