+++ srcs/vm/funcs/helper_functions/get_arg_value.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   get_arg_value.c                                    :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: orantane <orantane@student.hive.fi>        +#+  +:+       +#+        */
+/*   By: seronen <seronen@student.hive.fi>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/02/06 14:50:25 by Gerhard           #+#    #+#             */
-/*   Updated: 2021/02/09 15:36:22 by orantane         ###   ########.fr       */
+/*   Updated: 2021/02/09 20:32:59 by seronen          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,7 +27,7 @@ int      get_arg_value(t_carriage *carriage, int i, unsigned char *arena)
         arg_value = carriage->stmt->args[i];
     else 
     {
-        arg_idx_value = (short)carriage->stmt->args[i] % IDX_MOD;
+        arg_idx_value = carriage->stmt->args[i] % IDX_MOD;
         arg_value = read_bytes(arena, (carriage->abs_pos + arg_idx_value) % IDX_MOD, 4);
     }
     return (arg_value);
